refactor: drop errors package
